Delegate filtering logic to utils & PeopleFilters
